feat: implement audio engine deep module
