+++ docs/specs/effect-adoption-roadmap.html
@@ -621,7 +621,7 @@
 				<div class="metric-list">
 					<div class="metric">
 						<span class="badge good">Known</span>
-						<div>ABB has a strong boundary model and five current grey-box public APIs.</div>
+						<div>ABB has a strong boundary model and six current grey-box public APIs.</div>
 					</div>
 					<div class="metric">
 						<span class="badge good">Done</span>
