+++ src-tauri/src/audio/processor/AGENTS.md
@@ -5,6 +5,9 @@
 - Owns execution pipeline behavior in `src-tauri/src/audio/processor/`.
 - Source of truth for processor stage orchestration, cancellation checkpoints, and cleanup guarantees.
 - Output artifact commit policy lives in `src-tauri/src/output_artifact/`; processor finalization delegates final artifact decisions to that boundary.
+- This directory is the Audio Engine Deep Module private execution cluster. Code
+  outside `src-tauri/src/audio/` must use the parent `crate::audio` public strip
+  rather than importing processor files directly.
 
 ## Preferred Path
 
@@ -17,6 +20,9 @@
 
 - Finalization reports success only after the output artifact boundary returns final artifact truth.
 - Processor code must not directly perform final artifact `rename`, `copy`, or `hard_link`; `scripts/check-no-bridge-imports.sh` enforces this.
+- External FDK and native engine paths must use the shared finalization handoff;
+  adapter-specific code may stage media and write metadata, but final artifact
+  commit and success wording remain centralized.
 - Drop probe/inspection contexts before reopening the same path for decoder trials, processing, replacement, or another library.
 - Cancellation is checked at critical boundaries, including post-move and pre-success paths.
 - Terminal paths clean temporary resources to avoid residue across retries.
